refactor: Use assetLoader for audio file
