--- src/sound/pre-match-audio.js
@@ -1,13 +1,0 @@
-import beastAudioFile from 'assets/sounds/AncientBeast.ogg';
-
-export class PreMatchAudioPlayer {
-	constructor() {
-		this.beastAudio = new Audio(beastAudioFile);
-	}
-
-	playBeast() {
-		if (this.beastAudio.paused) {
-			this.beastAudio.play();
-		}
-	}
-}
